backend/requirements.txt
alembic==1.16.1
Flask==3.1.1
flask_cors==5.0.1
Flask_JWT_Extended==4.7.1
flask_limiter==3.12
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
marshmallow==4.0.0
marshmallow_sqlalchemy==1.4.2
protobuf==6.31.1
pytest==8.3.5
python-dotenv==1.1.0
python_bcrypt==0.3.2
SQLAlchemy==2.0.41
